RavenDB-24086 - skip heavy endpoints in test, move Authentication tests to stress tests
