+++ test/SlowTests/Tests/TestsInheritanceTests.cs
@@ -88,7 +88,7 @@ namespace SlowTests.Tests
                         select method;
 
             var array = types.ToArray();
-            const int numberToTolerate = 6391;
+            const int numberToTolerate = 6387;
             if (array.Length == numberToTolerate)
                 return;
